Moved TooltipHandler to widgets and all of the Screen classes to screens
