+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/api/appleskin/AppleSkinEventHandler.java
@@ -1,6 +1,6 @@
 package by.dragonsurvivalteam.dragonsurvival.api.appleskin;
 
-import by.dragonsurvivalteam.dragonsurvival.client.gui.ToolTipHandler;
+import by.dragonsurvivalteam.dragonsurvival.client.gui.widgets.ToolTipHandler;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateProvider;
 import net.minecraft.client.Minecraft;
 import net.neoforged.api.distmarker.Dist;
