+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/gui/screens/AbilityScreen.java
@@ -1,4 +1,4 @@
-package by.dragonsurvivalteam.dragonsurvival.client.gui;
+package by.dragonsurvivalteam.dragonsurvival.client.gui.screens;
 
 
 import by.dragonsurvivalteam.dragonsurvival.client.gui.widgets.buttons.*;
