+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/common/blocks/DragonAltarBlock.java
@@ -1,7 +1,7 @@
 package by.dragonsurvivalteam.dragonsurvival.common.blocks;
 
 
-import by.dragonsurvivalteam.dragonsurvival.client.gui.DragonAltarScreen;
+import by.dragonsurvivalteam.dragonsurvival.client.gui.screens.DragonAltarScreen;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateHandler;
 import by.dragonsurvivalteam.dragonsurvival.common.capability.DragonStateProvider;
 import by.dragonsurvivalteam.dragonsurvival.util.Functions;
